Update stubbed JSON to match new payload
This matches the new payload which now includes slots per location.

Dalston is now an example that can be used to test locations with a cut
off date for closure.

We now simply return the first entry in the stubbed JSON file as this
acts the same as a request to the booking location API does for a
request made with a child location UID.